\name{aTDT-method}
\docType{methods}

\alias{aTDT}
\alias{aTDT,FamilyExperiment-method}
\alias{aTDT,matrix-method}

\title{aTDT method}

\description{Foo}

%% \usage{

%% ## Constructors

%% FamilyExperiment(se, pedigree)
%% \S4method{FamilyExperiment}{SummarizedExperiment}(se, pedigree)
%% %\S4method{SummarizedExperiment}{missing}(assays, ...)
%% %\S4method{SummarizedExperiment}{list}(assays, ...)
%% %\S4method{SummarizedExperiment}{matrix}(assays, ...)

%% }
\author{Samuel G. Younkin <syounkin@jhsph.edu>}
