\name{CNVMatrix-method}
\docType{methods}

\alias{CNVMatrix}
\alias{CNVMatrix,GRanges,GRanges-method}

\title{CNVMatrix method}

\description{Foo}

%% \usage{

%% ## Constructors

%% FamilyExperiment(se, pedigree)
%% \S4method{FamilyExperiment}{SummarizedExperiment}(se, pedigree)
%% %\S4method{SummarizedExperiment}{missing}(assays, ...)
%% %\S4method{SummarizedExperiment}{list}(assays, ...)
%% %\S4method{SummarizedExperiment}{matrix}(assays, ...)

%% }
\author{Samuel G. Younkin <syounkin@jhsph.edu>}
