\name{SNPTrioExperiment}
\Rdversion{0.1}
\docType{class}
\alias{STE}
\title{Extension of SummarizedExperiment}
\description{
  Container for storing familial information for father, mother,
  offspring trios alongside the corresponding SummarizedExperiment object
}

\author{
Samuel G. Younkin
}

\keyword{SNPTrioExperiment}
