\name{trioClasses-package}
\alias{trioClasses-package}
\alias{trioClasses}
\docType{package}
\title{Classes and methods for Family data}
\description{Classes and methods for Family data}
\details{
  \tabular{ll}{
    Package: \tab trioClasses\cr
    Type: \tab Package\cr
    Version: \tab 1.0.6\cr
    Date: \tab Thu 30 May 2013 01:44:20 PM EDT\cr
    License: \tab THE BEER-WARE LICENSE\cr
  }
}
\author{
  Samuel G. Younkin <syounkin@jhsph.edu>
}
\references{
  ~~ Literature or other references for background information ~~
}
\keyword{ trioClasses }
